requirements.txt
openai>=1.0.0
pyyaml>=6.0
behave>=1.2.6
pytest>=7.0.0
pytest-asyncio>=0.26.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
loguru>=0.5.3
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.0.0
asyncio>=3.4.3
typing-extensions>=4.0.0